refactor: split render helpers into separate files
